feat: func isCacheable(Key) bool

+++ memcache4dalgo/transaction.go
@@ -11,6 +11,8 @@ import (
 type transaction struct {
 	ro dal.ReadTransaction
 	rw dal.ReadwriteTransaction
+	// isCacheable returns true if the key is cacheable
+	isCacheable func(key *dal.Key) bool
 }
 
 func (t transaction) ID() string {
@@ -22,11 +24,11 @@ func (t transaction) Options() dal.TransactionOptions {
 }
 
 func (t transaction) Get(ctx context.Context, record dal.Record) error {
-	return getRecord(ctx, record, t.ro.Get)
+	return getRecord(ctx, record, t.isCacheable, t.ro.Get)
 }
 
 func (t transaction) GetMulti(ctx context.Context, records []dal.Record) error {
-	return getMultiRecords(ctx, records, t.ro.GetMulti)
+	return getMultiRecords(ctx, records, t.isCacheable, t.ro.GetMulti)
 }
 
 func (t transaction) QueryReader(ctx context.Context, query dal.Query) (dal.Reader, error) {
@@ -38,7 +40,7 @@ func (t transaction) QueryAllRecords(ctx context.Context, query dal.Query) (reco
 }
 
 func (t transaction) Set(ctx context.Context, record dal.Record) error {
-	deleteCached(ctx, record.Key())
+	deleteCached(ctx, record.Key(), t.isCacheable)
 	return t.rw.Set(ctx, record)
 }
 
@@ -48,26 +50,29 @@ func (t transaction) SetMulti(ctx context.Context, records []dal.Record) error {
 }
 
 func (t transaction) Delete(ctx context.Context, key *dal.Key) error {
-	deleteCached(ctx, key)
+	deleteCached(ctx, key, t.isCacheable)
 	return t.rw.Delete(ctx, key)
 }
 
 func (t transaction) DeleteMulti(ctx context.Context, keys []*dal.Key) error {
-	deleteCachedByKeys(ctx, keys)
+	deleteCachedByKeys(ctx, keys, t.isCacheable)
 	return t.rw.DeleteMulti(ctx, keys)
 }
 
 func (t transaction) Update(ctx context.Context, key *dal.Key, updates []dal.Update, preconditions ...dal.Precondition) error {
-	deleteCached(ctx, key)
+	deleteCached(ctx, key, t.isCacheable)
 	return t.rw.Update(ctx, key, updates, preconditions...)
 }
 
 func (t transaction) UpdateMulti(ctx context.Context, keys []*dal.Key, updates []dal.Update, preconditions ...dal.Precondition) error {
-	deleteCachedByKeys(ctx, keys)
+	deleteCachedByKeys(ctx, keys, t.isCacheable)
 	return t.rw.UpdateMulti(ctx, keys, updates, preconditions...)
 }
 
-func deleteCached(ctx context.Context, key *dal.Key) {
+func deleteCached(ctx context.Context, key *dal.Key, isCacheable func(key *dal.Key) bool) {
+	if !isCacheable(key) {
+		return
+	}
 	mk := key.String()
 	_ = memcache.Delete(ctx, mk)
 	if Debugf != nil {
@@ -75,7 +80,7 @@ func deleteCached(ctx context.Context, key *dal.Key) {
 	}
 }
 
-func deleteCachedByKeys(ctx context.Context, keys []*dal.Key) {
+func deleteCachedByKeys(ctx context.Context, keys []*dal.Key, isCacheable func(key *dal.Key) bool) {
 	mks := make([]string, len(keys))
 	for i, k := range keys {
 		mks[i] = k.String()
